migrate from CRA to Next with tsx, also rename files and make project ready for component migrations
