+++ app/customer/components/Cart/CartItem.tsx
