+++ app/customer/components/Footer/Footer.tsx
