+++ app/customer/components/HomeCarousel/MainCarousel.tsx
