+++ app/customer/components/HomeSectionCarousel/HomeSectionCarousel.tsx
