+++ app/layout.tsx
@@ -1,0 +1,21 @@
+import type { Metadata } from 'next'
+import './index.css'
+
+export const metadata: Metadata = {
+    title: 'ShopSavvy',
+    description: 'ShopSavvy, an ecommerce website',
+}
+
+export default function RootLayout({
+    children,
+}: {
+    children: React.ReactNode
+}) {
+    return (
+        <html lang="en">
+            <body>
+                <div id="root">{children}</div>
+            </body>
+        </html>
+    )
+}
