+++ app/page.tsx
@@ -1,0 +1,3 @@
+export default function HomePage() {
+    return <p>heheheh</p>
+}
